feat: 服务层领域重构、API治理文档与增量功能
- 将 backend/app/services 拆分为 accounting/audit/auth/agent/doc_parsing/shared 等领域目录
- 新增印章识别、解析修正、Money 体系、文档标签等 API 与测试
- AGENTS.md 新增第九章 API 设计与边界原则
- 新增 api-boundary-governance-plan 及 requirements-domain-index 登记
- 安全与登录修复、CI 与前后端联调相关更新

未提交: backend/.env、qdrant 本地存储、一次性 fix 脚本与测试输出文件

+++ backend/app/schemas/accounting_entry.py
@@ -1,4 +1,6 @@
 from datetime import date, datetime
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
@@ -11,8 +13,8 @@ class AccountingEntryRead(BaseModel):
     summary: str | None
     account_code: str | None
     account_name: str | None
-    debit_amount: float
-    credit_amount: float
+    debit_amount: Decimal
+    credit_amount: Decimal
     counterparty: str | None
     normalized_text: str
     entry_line_no: int
